feat(web): injectable transport for research chat session hook
Extracts useConversationSession into conversationSession.ts with no
top-level import of @web/client or @web/wsHub, so callers that supply
their own transport (e.g. the pro web edition's host-routed research
slot) never statically pull in the app's singleton api client or
websocket hub. useChatSession.ts stays a thin wrapper that defaults to
the real client/subscribeChannel singletons, keeping chart/assistant
sessions and the default research session behavior unchanged.
